Improve GUI
GUI currently has no color background, text font and height is too high.

GUI require some improvements.

Let's change the background color by adding color to our dialogboxes and
add setfont to the dialogbox. Let's also adjust the height of the
dialogbox.

Setting the background color and the font. Adjusting the height in
screenbuilder is one of the better ways to improve gui.