Fix find feature and change a bit of duke personality
Find feature currently has a bug where it shows the found list even
though it cannot find any matches in the list. Duke talks in the default
way.

Need to fix the find feature so that it will not show the found list
when there is no matches. Change in duke personality as it is quite dull
.

Let's fix the find feature by setting isFind to false when there is no
matches. Let's change some text in the UI to make duke have more
personality.

Best way to fix this is shown as above.