Fix delete command
Current stage does not request for a new list after deletion.

Important to ask for a new list after some deletions as indexing will
change.

Let's fix the delete command by adding a condition if it isFind to
detect which lists to output and change the isDisplayList boolean to
false after finishing what it has to be done.

Best way to do it is the way shown above.