Add default tasks loading and jar bundling for storage
Previously, the storage file was placed in the root directory of the
project. This resulted in the storage not being bundled in the JAR file.

Storage has now been moved into resources. In addition, a default tasks
storage has been added to load in some default, initial tasks when the
app is first launched.