Add user guide, fix tasks adding bug and add Ui picture
Currently, there is no User guide and adding tasks will not ask for
display of new list. There is no screenshot of the Ui in the docs too.

Need to fix all these for better user use.

Let's update the user guide, add the Ui screenshot and fix the bug.

Only way to do it.