Implement C-Help: Give help to users. Provide in-App guidance to users. Add a command to access help page.
New users may not know how to use the app. If they use 3 invalid
commands, a help message will be generated, and a command to access
the command list will be given.

With the addition of command list, accessibility will improve and new
users can learn how to use the app easily.