Fix command body parsing
The user input "deadline /by" fails to extract the empty command body,
and the 'by' parameter is improperly parsed due to the empty body.

It needs to be fixed so that Duke can detect the empty body, and give
the proper error message.

Let's update the Regex that's used to split the body from the
parameters, excluding the space before the param slash '/'.

By splitting on the slash, it would properly obtain the empty body,
should the body be empty.