Implement C-Help
New users using the app are not familiar with the commands to use.

Adding a "help" command allows users to familiarize themselves with the
commands. The help message is also shown when Genie first launches.

First-time users are immediately given guidance, and new users are given
guidance at any point of time using the "help" command.