Add assertions for tasks
There are checks to ensure the capacity of task lists are not violated.

However there is currently no easy way to know when a task list is overloaded past its capacity.

Adding assertions will help with enforcing the invariant that the number of elements in the task list should not exceed the capacity.