Refactor start method
Main start method for JavaFX currently has two layers of abstraction.

This does not provide a good high level overview of what the method is supposed to do, considering how high level the start method is.

Extracting and grouping parts of the start function into new methods helps to reduce the layers of abstraction to one, and makes the method much easier to read.