Add assertions
The current Command classes that are responsible for handling commands
that involve creating a new task, assumes that a Task object will be
created upon execution. These classes are:
* DeadlineCommand
* EventCommand
* TodoCommand

These assumptions will have to hold in order for these classes to work
as intended. However, if they do not hold, it will result in errors in
other areas of the code.

Having assertions will ensure that the original assumption holds in the
respective Command classes.