Add assert statements
Adding assertions allows us to be sure that the correct flow of
code is happening during runtime.

Let's add assertions to code statements that check whether the split
operation really resulted in split of string into 2. Moreover,
checking if resources are null for images.

All of these prevents exceptions and errors due to programmer error
to occur during runtime.