Improve and customize GUI: Change background and increase padding on chat boxes
The current GUI of the app looks dull and uninviting. Changing the
background color and increasing the padding on chat boxes can enhance
the overall look and feel of the app.

In this commit, I have made the following changes:

Changed the background color to a more vibrant and appealing one
Increased the padding on chat boxes to make them look more spaced out
and prominent.
These changes will make the app look more modern and engaging, which
will likely increase user engagement and satisfaction.