delete original Main class
The functions of the Main class is substituted, so delete it to make the code cleaner.

deleting Main as a whole class, while naming the new main class Main2 before the deletion

It ensures the code has basic functions before new structures are fully implemented.

uploaded the corresponding jar file after this.