Update README.md
README.md does not contain instructions that tells the user how to use
the bot.

Update README.md with instructions in GFMD format on what each command
does and what format should the use input for the commands to work.

Instructions written in GFMD format for better visualization and
formatting