Duke Class and Storage class: add Assertions
Some assumptions are made when the program is executed, checks by assertions are needed so that the runtime can verify them.

Let's check the size of the arraylist at several parts when new tasks are added.

When a new task has been successfully created, the array of task should contain more than one task.