Added Javadoc, Matched coding standard
Added Javadoc for classes and methods without Javadoc.

Removed unnecessary throwing statement of DukeException in execute method of some commands.

Refactored class name of TodoCommand to be ToDoCommand so that it follows the coding standard.

Changed the names of add method in TaskList so that it is easier to identify which task is being added to the TaskList.

Modified find method in TaskList to show the index to match the index of the task in the whole TaskList, not in the list of matching tasks.