Removes deprecated Ui class from Duke
Duke currently retains the Ui class from its previous functions as a
CLI chat bot.

However, as new updates are given to the GUI form of Duke, the CLI form
is no longer maintained as well. The CLI form now lacks the Undo function
that is already implemented for the GUI.

Let's delete deprecated functions that still use Ui.

This will reduce the errors and is part of the deprecation process of Ui.