Update Storage to use assertions. Storage methods use exceptions to handle assumptions - files existing.
The use of exceptions makes it seem like the file not existing is a
condition that could potentially occur.

Extracting this code into assertions makes the code more intuitive as
this behaviour is assumed to be true.

Let's add assertions in the code in places where it is assumed that the
file exists, as it should have been created if it did not.