Update all Commands to state assertions. Commands are assumed to be initialised with non-null inputs.
There is no explicit catch for this, and it is possible that one may
mistakenly use null in their declaration of a command.

Let's add an assertion that the input is not null when initialising the
command.

It is more intuitively stated so that the class is not misused, and if
it is used wrongly, java can catch that mistake.