New class Update: Implement update dates functionality
The bot does not allow users to update event time. This posts potential
inconveniences to the users. Firstly, if users keyed in the wrong date,
they have to delete the task and enter a new one. Secondly, if the user
decides to postpone the task, they have to recreate the task as well.

Let's implement a new function to allow users to update the dates of
an existing task.

Refer to this requirements for the specific requirements:
https://nus-cs2103-ay2223s2.github.io/website/se-book-adapted/
projectDuke/index.html#level-4-todos-events-deadlines