Update Functionality: Enables wrong data format detection
The bot does not detect the wrong user format for the update command,
it crashes when user inputs random strings after the update keyword.
This disrupts the user experience, and leave the user not knowing what
is happening and what is supposedly the correct format.

Let's implement a user input validation function that tells the users
when they key in the wrong data format for the update command.

Refer to this requirements for the specific requirements:
https://nus-cs2103-ay2223s2.github.io/website/se-book-adapted/
projectDuke/index.html#level-4-todos-events-deadlines