Add assertions
No way of checking If the commands checking correct input function runs or not .
Later on if the code is changed , invalid argument could be sent to the run function of the commands.
Let's add assertions to check if valid arguments are being passed i.e. checking this.correctInput() function runs.
Assertion terminate indicating invalid argument.