Update Storage class to automatically create directory
The Storage class was failing to create a directory automatically when
the jar file was running, resulting in errors when trying to save files
to non-existent directories.

To fix this, the code has been modified to check if the directory exists
before creating a file. If the directory does not exist, the code will
now automatically create it before creating the file.

This change makes the Storage class more robust and user-friendly. Users
will no longer need to manually create directories before running the
jar file.