Add assertions
The Storage class handles the reading, writing and parsing of the storage text file and handles corrupt storage with exception handling.

The Storage class can use some assert statement handling.

Let's add some assert statements where possible.

This will improve the QOL of the code and it will be happier.