Add assert statement in the TaskList class
Assertions are used to define assumptions about the program state
so that the runtime can verify them.

Adding assertions in the code statements so that it can check whether
the split operations used in the input string result in the correct
output.

It can help to detect the mistakes in the codes.