Modify the codes based on the code quality
Some of the classes had some code quality violation that needed to
be fixed according to the code quality rules.

Deleting dead code and unused codes so that not trip up readers.

Avoiding deep nesting and complicated expressions so that much more
easier for reading.

Trying to follow the single level of abstraction principle.

These are helping to run-time efficiency and understandability.