Correct some outputs to match the expected output
Some tests are not passing due to change in output format.

Output format needs to consistent for better user experience.

Changing the output format of date in deadline and event to be more descriptive and consistent with the expected output.

Descriptive date in output will make it easier for user to read.