Fix load .txt file bug and improve UX
Contents of .txt file disappears after reading.
Users with existing saved data cannot view their list when launching
the app.

Showing the list upon launching the app serves as a refresher for
existing users. When there're no contents in the .txt file, assume that
the user is new and shows the help message instead.