Find command: make matching case insensitive and matches to dates
Find functionality is case sensitive, and only matches to tasks description.

A case insensitive find is more user-friendly because users cannot be expected to remember exact case of keywords.
A find feature that matches dates is more user-friendly for users to see what events and deadlines they have on that day.

Let's,
* update the search algorithm to use case-insensitive matching
* update the search algorithm to match dates in the form <yyyy-MM-dd> or <MMM dd yyyy>