Command classes: extract a parent class Command
Command classes are all user commands parsed from user inputs, and
are executable. Extracting a Command class allows for better OOP
and modularity of the code.

Execution of commands is extracted from Parser class into Command
classes. Each command is a subclass of the abstract Command class
and each command implements execute() differently.

This allows for modularity and better OOP so more commands can easily
be added in the future. The code is also easier to understand and
functionality and execution of the commands are abstracted away
into the corresponding Command classes instead of just the Parser class.