Add assertions to Duke
There are no assertions to document important assumptions that should
hold at various points in the code.

Need to document important assumptions that should hold at
various points in the code so that the runtime can verify them.

Add assertions to Duke.java, Storage.java and TaskList.java.

This is done because an assertion failure indicates a possible bug and
the sooner the execution stops, the safer it is.