Add assert statements at various points in the code
Assert statements were added to document important assumptions that
should hold at various points in the code.

Let's add assert statements in the following,
* In the DEADLINE command to check that the deadline description
contains a "/by" or "(by:" field.
* In the EVENT command to check that the event description contains a
"/from" or "(from:" field.
* In the EVENT command to check that the event description contains a
"/to" or "to:" field.
* In the TODO command to check that the todo description is not empty.
* In the add(Task task) method to check that the task parameter is not
null.
* In the mark(int index) method to check that the index parameter is
within the bounds of the list.
* In the unmark(int index) method to check that the index parameter is
within the bounds of the list.
* In the delete(int index) method to check that the index parameter is
within the bounds of the list.
In the getTask	(int index) method to check that the index parameter is
within the bounds of the list.