Add sort feature
Previously list command simply lists down the tasks in the order
that they were added to the list. This could be messy for the user.

As such, the sort feature allows the list to be sorted in
chronological order according to the dates. Todo tasks without
date would be listed at the very top of the list.