Storage Loading: enable loading of previous list
Due to changes made in A-MoreOOP, issues were generated for loading previous list.

Adjustment made to parse date value in stored previous list

Previous list dates are stored in "MMM dd yyyy, DAY_OF_WEEK, hh:mmAM/PM" rather than in "yyyy-MMM-ddTHH:mm", which is the required format of users when entering data.

This parsing method could be used to allow users to enter dates in the format that is stored in the lists, rather than in the current required format.  Might make this available in future commits.