Separate Gui from the main logic of duke
The gui is mixed with duke's logic of handling input in the Duke class.

This way of implementation makes it harder to further extend the Gui,
and makes the overall logic much more difficult to follow.

Extract the Gui portion in different classes, i.e. DialogBox,
MainWindow.

This is to allow developer to see the overview of code structure in
an easier way.