feat: support creation of tasks with places
As per D-Places extension specification, we add the ability
to add a place to any item, and create a task with this place

First of all, Task must be modified to use String instead of Optional<String>
since we want it to be Serializable. We convert it to and from Optional<String>
as necessary.

Next, support for adding a task with a place is added via an optional /place
parameter to event, task and todo commands.