Add GUI using JavaFX and FXML
Our current text-based UI is plain and has no personality.

Adding a simple GUI using JavaFX can greatly increase the user
experience and add some personality to the text-bot.

JavaFX is used because of its cross-platform capabilities and ease of
implementing.

Refer to this tutorial to show how JavaFX was implemented:
https://se-education.org/guides/tutorials/javaFxPart1.html