Clean up some code.
Need to make another commit so the A-FullCommitMessage tag will appear.
Currently its not appearing because i don't have an empty commit with no existing tags.
All of my commits have tags on them so I need to make a new commit so I can put
the A-FullCommitMessage Tag. Thank you!