Refactor code for GUI
The current parser implementation and UI is for a text-based UI.

After switching to a GUI, a code refactoring is necessary to print the
correct inputs to the GUI.

Parser methods will be changed to return string values instead of
console prints.

This is so that String values can be printed to the GUI instead of
console print statements.