add assertions
When TaskList methods are called, the ArrayList of Tasks should never be
null.

Adding assertions allows us to ensure that this invariant is always
true.

Let's add assertions to the TaskList methods to ensure that all TaskList
Objects which are targets for method invocation have non-null ArrayLists
at the time of method calling.