Add data file corruption error handling
The program does not take care of the case where the
data file used has been corrupted.

This will lead the program to crash if the data file
is corrupted.

Let's change the program to handle this by making it
delete all data in the file without a warning, and start
afresh, if the file is detected to be corrupted.

This will help avoid the program from crashing in the
event of a data file corruption.