Add more error handling
MarkCommand does not handle the scenario where a task number does not
exist in the TaskList. Todo, Event and Deadline also does not handle
scenarios where task details are empty or invalid.

When a user enters an input that belongs to such scenarios, they may
encounter runtime errors which are undesirable.

The respective classes can now handle errorneous inputs such as:
* marking/unmarking a task number that does not exist
* wrong command format
* empty task descriptons