Fix bug in reading empty files
A NullPointerException is being thrown when the file where
tasks are stored ("duke.txt") is exists but is empty.

This wrongly terminates the execution of the program due
to some resources not being initialized.

Let's try to fix that by handling the DukeException that
gets thrown when there are no saved tasks to load, and
re-initializing resources to ensure that the application
runs.