Refactor loadTaskList() method
The loadTaskList() method is long and uses different layers of abstraction.

The code does not follow the SLAP principle. Refactoring the code with smaller methods also improves readability. The behaviour when there is an error with the data file is not made explicitly clear.

Let's,
* extract parseTaskList() and parseTask() methods from loadTaskList()
* extract clearing of data file into its own method

Doing this allows us to:
* understand better how the data file is processed
* change the behaviour of clearDataFile() method easily