Add a Sort feature
To give proper ordering to the list of tasks, a sort method should be
added to allow the user to sort the tasks in the list.

Sorting the tasks in alphabetical order according to the task description
will enable the user to better locate tasks in the list.

To implement this feature, let's:
* Implement the Comparable interface in the Task class and override the
compareTo() method to compare task descriptions.
* Add a "Sort" command for the user to invoke the sort feature.