build.gradle: Change mainClassName to "duke.Launcher"
mainClassName in build.gradle does not match with the actual main class
in the current project. JAR files generated using the command
"./gradlew shadowJar" are therefore unable to find the main class.

Let's change mainClassName to "duke.Launcher" as it is the actual entry
point of our program.

This will allow JAR files built using gradle to run properly.