Add sort by deadlines feature
The tasks are stored in the order in which they are created.

It is difficult for the user to quickly see which deadlines are due next.

Let's add a new command to sort the list of tasks chronologically based on the due date of the deadlines and preserve the order otherwise.

This change improves the usability of the chatbot application by allowing users to easily view and manage their tasks in order of their deadlines.