Changed name of the software, Fixed minor bugs
Renamed the software to DOggi, changed responses shown to the user to meet the concept.

Fixed the bug that does not show the full content when the content is long.

Added greeting message when the software starts.

Added execution of the software when EndCommand is executed.

Added more condition for commands involving index so that index smaller than 1 (out of range) throws an exception.