add sort functionality, empty list exception
The user may wish to sort the task list by description to make it easy to
read.

Let's add a sorting feature to enable the user to do this.

Some commands such as list and sort will access the list elements. If
the list is empty, they are unable to work as intended.

Let's add an exception to show the case when an empty list is being
accessed by commands that work with the whole list.