Fix checkstyle errors, add JavaDocs and cite sources
Checkstyle errors were introduced during GUI development.
These were fixed, and JavaDocs were added or updated for the modified classes.

During development, StackOverflow snippets were used a few times without attribution,
these were fixed by adding a comment stating the author's name, date the answer was written,
and a link to the answer along with the name of the license under which the answer is provided.