Commit message for Level-9
Add Find command

As the program expands, the user is now able to add more tasks that they want.

In the case where the list of task gets too long, it might be difficult for
the user to find a specific task that they want to view or mark.

Therefore a find command is important as it allows the user
to filter the task list using keywords, allowing them to find their
tasks much faster.