Commit message for Level-10
Add GUI and JavaFX support

As the program grows, text based UI is unable to fully and easily allow
the user to view and edit information.

Therefore a GUI is important as it allows users to view their information,
as well as input data much easily.

It also allows easier debugging as there is now visual indication for any bugs.