Add screenshot showing product Ui Add user guide
Without user guide or product ui screenshot,
users will not know what to except
and will not know how to use the program.

Adding a product Ui screenshot and a user guide ensures that users
are able to learn how to use the application easily, and encourages
users to use the application.

Let's
* add an image that showcases the product's Ui
* add a comprehensive user guide