Abstract getResponse() method body
The getResponse() method body does not adhere to SLAP and has an arrowhead style code. This makes the code messier and harder to debug.

Refactoring the method body will improve code quality.

Let's create a Command class which handles the execution of the user input, and split the code into their respective methods to mitigate the arrowhead style code.