Add assertions
The code does not use assertions.

It is difficult identify bugs. The location of compilation errors may
not be explicitly stated by the compiler, and for logical bugs it is not
easy to debug without knowing which variables or methods are bugged.

Let’s add `assert` statements at various points in the code.

An Assertion is a statement that tests assumptions about the code. It
can help the programmer to read the code better, detect any defects.
The assert statement halts the code at the exact point if it is not
fulfilled, allowing the programmer to identify the exact location there
is a bug. This makes the process of debugging easier.