Implement Loan extension
There is no way to keep track of tasks which involve keeping track of
money owed to someone.

Let’s implement a class Loan which inherits Task, and the methods
associated to the Loan class.

A Loan contains the following fields:
* Name of the task
* Borrower
* Lender
* Amount
* Status

With the Loan class, the user can keep track of loans, including namely
who owes how much money, and to who.

Some of the associated methods include:
* Loan.createLoan() which creates a new Loan based on user input
* Loan.createLoanFromLog() which loads a Loan from a save