Add a way to archive items
A way to archive every task and action is added. Tasks archived will be
saved in archive.txt in the folder "data". Deleted tasks will be
removed from tasks.txt but remains stored in archive.txt. User can use
command "archive" to archive a copy of the current tasks in list. This
will be saved in archiveAll.txt in folder "data". archive.txt will still
contain history of all tasks, not just tasks in current list.