Cluck! now working in GUI
Previously Cluck!'s GUI merely repeated the user input, now it has been fixed such that Cluck! actually takes in the user input and returns a response after executing the command.

This was done by changing the getResponse method in Cluck class which takes in the user input as a String argument and parses it into a command, which is then executed and returns the response from executing the command.