Edit to improve code quality in Cluck class
Previously Cluck contained high levels of nesting.

Additional methods have been added to improve readability and overall quality of code in the run() method in cluck.
getResponse() is called to get the response from a userInput. getResponse calls two other methods:
getCommand - takes in user input and parses it into a command
executeCommand - takes in a command and executes it, returning the response after executing the command