Added flexibility for user date tiem inputs
Previously the date time input of the user has to be precise, for example if the user wants a deadline in march they would have to input "Mar" exactly.

This commit changes it such that the date time formatter is no longer case sensitive, and the user can even enter the numerical equivalent of the month they desire.