Fixed date time formatting for tasks
Previous implementation to allow for more flexble parsing of date time given by the user cause some problems. In the previous commit I added optional patterns to allow for user to input various forms of date time, e.g. the user can input either 06, Jun, or June for a date on june. The user also has the option to omit the century of the year (i.e. the "20" in "2023") or minute of the time.

The change in FORMATTER in the task class cause the toString() and makeSaveFormat() methods in Event and Deadline classes to format the LocalDateTime classes into the wrong format, as the FORMATTER's optional patterns are all used to format the LocalDateTime, resulting in strange date strings like "21 03MarMarc 202323 1200".

To resolve the issue, I added an additional formatter by the name of default formatter where the DateTimeFormatter does not have optional patterns, and thus gives us consistent and correct string formats of the date when used by toString() and makeSaveFormat() methods.